requirements.txt
requests
dnspython
PyYAML
